add payment entity

+++ payment/input.go
@@ -1,6 +1,6 @@
 package payment
 
 type InputPayment struct {
-	ProductId int64   `json:"product_id" building:"required"`
+	ProductId int     `json:"product_id" building:"required"`
 	PricePaid float64 `json:"price_paid" building:"required"`
 }
